reafacto(contributions): cleanup + tests
